requirements.txt
# study1 Streamlit 앱 의존성 (루트에서 pip install -r requirements.txt 용)
# 상세: study1/requirements.txt (내용 동일)
streamlit>=1.33.0
Pillow>=10.0.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0

# API providers (install the one you plan to use)
# pip install anthropic
anthropic>=0.39.0
# pip install google-generativeai
google-generativeai>=0.8.0

# Google Drive 업로드 (대화 로그 저장)
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.2.0
